requirements.txt
pandas==1.1.4
rank_bm25==0.2.1
spacy==2.3.2
tqdm==4.51.0
pyspellchecker==0.5.5
